scheduler_functions/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-cosmos
langchain==0.1.12
langchain-openai==0.0.8
jsonschema==4.21.1